Move API entities to api package
