+++ app/src/main/java/com/rebalance/ui/component/main/GroupSpendingList.kt
@@ -34,7 +34,7 @@ import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.rebalance.PreferencesData
-import com.rebalance.backend.entities.Expense
+import com.rebalance.backend.api.entities.Expense
 import com.rebalance.backend.service.BackendService
 import com.rebalance.util.alertUser
 import compose.icons.EvaIcons
